sonar

See merge request petemarakos/estate-fe!437

+++ src/pages/api/calendar/[userId]/events/removeAll.page.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next/types";
 import calendarService from "../../_service/CalendarService";
-import toNumber from "@/utils/toNumber";
+import toNumberSafe from "@/utils/toNumberSafe";
 
 export default async function handler(
     req: NextApiRequest,
@@ -18,7 +18,8 @@ export default async function handler(
 
         const { date } = req.body;
 
-        const iUserId = toNumber(userId);
+        const iUserId = toNumberSafe(userId);
+        if (iUserId === -1) throw new Error("Bad userId");
 
         await calendarService.deleteDayEvents(iUserId, date);
